Decomposed TagsEditFormAutocomplete
